projeto_analise/requirements.txt
python-dotenv
google-genai
anthropic
streamlit
plotly
markdown2
xhtml2pdf
pymupdf
pytest
